refactor: separate responsibility about downloading files from slack
- add files:read scope on slack user token
